Nodes data structure

+++ src/linking_matrix.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void GetLinkMatrix(double *link_mat,int Ne,int order)
+void GetLinkMatrix(int *link_mat,int Ne,int order)
 {
 	int p = order+1;
 	for(int i=0; i<Ne; i++)
